backend/requirements.txt
# Backend dependencies — install: pip install -r requirements.txt

# --- Core ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
passlib>=1.7.4
PyJWT>=2.8.0
pypdf>=5.0.0
PyMuPDF>=1.24.0

# --- Optional: OCR for scanned PDFs ---
# Install system packages first (e.g. Tesseract + Poppler), then uncomment and install:
#   pdf2image>=1.17.0
#   pytesseract>=0.3.10
#   Pillow>=10.0.0
# Set RESUME_OCR=1 in backend/.env to enable (see pdf_text.py).
